Hide system bars to see full theme in list
And drop fragments in favor of traditional activities
because that's simpler and better suited for this kind
of architecture. Sure, Fragments allow us to separate
UI components from Activity objects, but that's also
more complex and this is a small app.

Hiding the system UI (status and navigation bar) works
well for devices *without* display cutouts now. On devices
with cutouts the view is jumping because the status bar
slides out but immediately snaps back. This can be fixed
by setting `android:windowLayoutInDisplayCutoutMode` to
`shortEdges` but that breaks playing the FAB above the
navigation bar :(

Apparently, when `shortEdges` is set, `fitsSystemWindows`
doesn't work anymore.