Make gallery and editor toolbar collapsable
This is way more complex than it should be.

To make the Toolbar for the Gallery- and EditorActivity
collapsable, it's required to have an AppBarLayout around
the Toolbar what means it's required to have an extra,
explicit Toolbar layout.

But this explict Toolbar layout doesn't slide out as nicely
as built-in from Theme.AppCompat. It just fades out and
doesn't move. This is probably because the explicit Toolbar
isn't the first child in the layout but with the theme in
the background, it can't be.

So two AppTheme's are required: Theme.AppCompat with the
built-in ActionBar for ThemeActivity to make the full screen
animation look good. And Theme.AppCompat.NoActionBar for
Gallery- and EditorActivity to make the Toolbar collapsable.

If you look closely, you'll notice that the up arrow in
EditorActivity has a slightly different color than the one
in ThemeActivity - obviousely Android chooses different
settings for the built-in ActionBar and for the explicit
Toolbar. Oh the joy.