Add a add-theme-FAB to the GalleryActivity
And remove it from the action menu in that activity.

Makes adding themes somewhat inconsistent because it
remains to be an action item in the ThemeActivity.

On the other hand, a FAB can be tapped more easily
and there should be a dominant default action in the
GalleryActivity as well.