Replace thumbnail image with a vector drawable
This doesn't make a difference for the system live wallpaper picker
where this thumbnail is used, but it keeps Samsung's One UI 6.0 from
using the thumbnail for the background of the app drawer (and other
places like the task switcher).

Apparently, in One UI, when a live wallpaper is set, the thumbnail of
the live wallpaper is heavily blurred and used as a static background
for the app drawer. This results in a rather puzzling, out of context
background instead of the live wallpaper, which is quite annoying.

Fortunately, One UI can't process a vector drawable so we can use
that to get the live wallpaper back in these situations.